added updated structure with auth, database and model
