+++ .env
@@ -1,0 +1,2 @@
+SECRET_KEY="AGENTGAMES2024"
+ADMIN_PASSWORD="BOSSMAN"
