+++ .gitignore
@@ -10,6 +10,15 @@ htmlcov/
 # pytest cache
 .pytest_cache/
 
+# Environments
+.venv
+env/
+venv/
+ENV/
+env.bak/
+venv.bak/
+agent_games/
+
 # pip cache
 Pipfile.lock
